+++ changelog.md
@@ -4,5 +4,6 @@ Changelog
 0.4.0: ( Apr. 4, 2018)
 ----------------------
 
-* adds command parameters;
+* adds command parameters (param value list, named param list);
+
 * adds command group;